requirements.txt
langchain==0.3.1
openai==1.51.0
python-dotenv==1.0.1
langchain_community==0.3.1
duckduckgo-search==6.2.13
#wikipedia==1.4.0
langchain-openai==0.2.1
playwright==1.47.0
pydantic-core==2.23.4